% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildIndex.R
\name{buildIndex}
\alias{buildIndex}
\title{Build an index for querying by variant ID from an LDZipMatrix object}
\usage{
buildIndex(ld, variant_file)
}
\arguments{
\item{ld}{An external pointer to an \code{LDZipMatrix} object,
typically created with \code{LDZipMatrix()}.}

\item{variant_file}{optional path to the variant annotation file (e.g. a 
\code{.pvar} file) containing the variant information 
(\code{CHROM}, \code{POS}, \code{ID}, \code{REF}, \code{ALT}). A standard PLINK2 \code{.pvar} file works here.}
}
\value{
\code{NULL}. The index is written to the SQLite database (\code{$prefix.sqlite})
}
\description{
This function builds an index on variant identifiers from the 
associated \code{.vars.txt} file of the \code{LDZipMatrix} object. 
Optionally one can pass their own variant file.
}
\details{
The index allows efficient mapping from variant IDs to internal 
integer indices, which can then be used with \code{fetchLD()} 
or \code{getNeighbors()}.
}
\examples{
\dontrun{
# Open an LDZip matrix
prefix <- file.path(system.file("extdata", package = "LDZipMatrix"), "g1k.chr22.ldzip")
ld <- LDZipMatrix(prefix)

# Build an index
idx <- buildIndex(ld)

# Fetch LD values by variant IDs
fetchLD(ld, "rs587755077", c("rs587631919", "rs587661542"), type = "PHASED_R")
}

}
