% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchLD.R
\name{fetchLD}
\alias{fetchLD}
\title{Fetch LD values from an LDZipMatrix object}
\usage{
fetchLD(
  ld,
  row,
  col,
  types = c("UNPHASED_R"),
  pairwise = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{ld}{An external pointer to an \code{LDZipMatrix} object,
typically created with \code{LDZipMatrix()}.}

\item{row, col}{An Integer (1-based) or character scalar/vector of variant indices/IDs
specifying the row(s)/column(s) to query.}

\item{types}{a character scalar/vector of types of linkage statistics to 
return. Choices are \code{"PHASED_R"}, \code{"UNPHASED_R"}, \code{"PHASED_R2"}, \code{"UNPHASED_R2"}, \code{"D"}, \code{"DPRIME"}.}

\item{pairwise}{(NOT implemented yet) Logical flag controlling the return type:
\itemize{
  \item \code{TRUE}: return pairwise LD values, where each element of
    \code{row} is matched with the corresponding element of \code{col}
    (both must have equal length).
  \item \code{FALSE}: return the full LD submatrix spanning all
    combinations of \code{row} and \code{col}.
}}

\item{simplify}{Logical flag. If \code{TRUE}, the return value is simplified
according to the shape of the query and the number of statistics:
\itemize{
  \item Single statistic, (\code{1 x 1}), (\code{1 x N}) or (\code{N x 1}): a \code{vector}.
  \item Single statistic, (\code{N x N}): a \code{matrix}.
  \item Multiple statistics, (\code{1 x 1}), (\code{1 x N}) or (\code{N x 1}): a \code{data.frame}.
  \item Multiple statistics, (\code{N x N}): a \code{list} of \code{matrix}.
}
If \code{FALSE}, results are always returned as a list of matrices, one per statistic.}
}
\value{
\itemize{
    \item A numeric vector if either \code{row} or \code{col} is a vector
      (the other scalar), and only one statistic is requested.   
    \item A numeric matrix if both \code{row} and \code{col} are vectors and
      a single statistic is requested.
    \item A \code{data.frame} if either \code{row} or \code{col} is a vector
      (the other scalar), and multiple statistics are requested.   
    \item A list of matrices if both \code{row} and \code{col} are vectors
      and multiple statistics are requested (or if \code{simplify=FALSE})
  }
}
\description{
This function retrieves linkage disequilibrium (LD) values from a
compressed \code{LDZipMatrix} object. Queries can be made either by
integer indices (1-based) or by variant identifiers (IDs). If both
indices and variant IDs are provided, variant IDs take priority.
}
\note{
\itemize{
  \item In order to use variant identifiers (instead of indices), please index the \code{LDZipMatrix} object using \code{buildIndex(ld)}
  \item Variants in the returned object are sorted in order of their variant indices, and not the same order as the input in \code{row} or \code{col}
}
}
\examples{
\dontrun{
# Open an LDZip matrix
prefix <- file.path(system.file("extdata", package = "LDZipMatrix"), "g1k.chr22.ldzip")
ld <- LDZipMatrix(prefix)

# 1 x 1 query, single statistic → scalar
fetchLD(ld, 2, 191, type = "PHASED_R")

 # 1 x 1 query, multiple statistics → data.frame
fetchLD(ld, 2, 191, type = c("PHASED_R", "DPRIME"))

# 1 x N query, single statistic → vector
fetchLD(ld, 2, c(2,129,191,307), type = "PHASED_R")

# 1 x N query, multiple statistics → data.frame
fetchLD(ld, 2, c(2,129,191,307), type =  c("PHASED_R", "DPRIME"))

# N x N query, single statistic → matrix
fetchLD(ld, 1:5, 1:5, type = "PHASED_R")

# N x N query, multiple statistics → list of matrices
fetchLD(ld, 1:5, 1:5, type = c("PHASED_R", "DPRIME"))

## Query with variant identifiers
# Build index (if it doesn't exist already)
buildIndex(ld)

# 1 x 1 query, single statistic → scalar
fetchLD(ld, "rs587755077", "rs587631919", type = "PHASED_R")

# Define a small set of variant IDs
vars <- c("rs587725733", "rs587631919", "rs587661542", "rs587718290", "rs587729333")

# 1 x N query, single statistic → vector
fetchLD(ld, "rs587755077", vars, type = "PHASED_R")

# 1 x N query, multiple statistics → data.frame
fetchLD(ld, "rs587755077", vars, type = c("PHASED_R", "DPRIME"))

# N x N query, single statistic → matrix
fetchLD(ld, vars, vars, type = "PHASED_R")

# N x N query, multiple statistics → list of matrices
fetchLD(ld, vars, vars, type = c("PHASED_R", "DPRIME"))
}

}
