% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeighbors.R
\name{getNeighbors}
\alias{getNeighbors}
\title{Find neighboring variants in an LDZipMatrix}
\usage{
getNeighbors(
  ld,
  variant,
  type = "UNPHASED_R2",
  abs_threshold = 0.8,
  genomic_length = 1e+05
)
}
\arguments{
\item{ld}{An external pointer to an \code{LDZipMatrix} object,
typically created with \code{LDZipMatrix()}.}

\item{variant}{Integer (1-based)/Character scalar referring to the reference variant index/identifier around
which neighbors are sought.}

\item{type}{Character scalar referring to the type of linkage statistic sought.}

\item{genomic_length}{Integer scalar referring to the window size in base pairs
(e.g., 1e6 for +/-1 Mb).}
}
\value{
Integer/Character vector of variant indices/indentifiers. 
\itemize{
\item If \code{variant} is a variant index (\code{Integer}), the function returns an integer vector (1-based) of variant indices
\item If \code{variant} is a variant identifier (\code{Character}), the function returns a character vector of variant identifiers
}
}
\description{
This function finds all variants exceeding a specified LD threshold
for a given statistics type, within a given genomic window around 
a query variant, using an \code{LDZipMatrix} object.
}
\note{
\itemize{
  \item In order to use variant identifiers (instead of indices), please index the \code{LDZipMatrix} object using \code{buildIndex(ld)}
}
}
\examples{
\dontrun{

# Open an LDZip matrix
prefix <- file.path(system.file("extdata", package = "LDZipMatrix"), "g1k.chr22.ldzip")
ld <- LDZipMatrix(prefix)

# Build index (if it doesn't exist already)
buildIndex(ld)

# Find neighbors with r^2 > 0.8 within +/-500kb of rs587755077
getNeighbors(ld, "rs587755077", type="PHASED_R", abs_threshold = sqrt(0.8), genomic_length=500000)


# Find neighbors within +/-100Kb of rs587755077 that are in perfect LD (D'=1)
getNeighbors(ld, "rs587755077", type="DPRIME", abs_threshold = 1, genomic_length=100000)

# Verify their D' value
res = getNeighbors(ld, "rs587755077", type="DPRIME", abs_threshold = 1.0, genomic_length=100000)
fetchLD(ld, "rs587755077", res, "DPRIME") 

}

}
